+++ src/ckanapi_harvesters/builder/mapper_datastore_multi.py
@@ -90,12 +90,8 @@ class RequestMapperABC(DataSchemeConversion, ABC):
         for file_query in self.download_file_query_list(ckan=ckan, resource_id=resource_id):
             yield file_query
 
-    def download_file_query(self, ckan: CkanApi, resource_id: str, file_query:dict) -> pd.DataFrame:
-        return ckan.datastore_search(resource_id=resource_id, **file_query, search_all=True)
-
-    # TODO: expose generator to enable writing to files in append mode:
-    # def download_file_query(self, ckan: CkanApi, resource_id: str, file_query:dict) -> Generator[pd.DataFrame, Any, None]:
-    #     return ckan.datastore_search_generator(resource_id=resource_id, **file_query, search_all=True)
+    def download_file_query(self, ckan: CkanApi, resource_id: str, file_query:dict) -> Generator[pd.DataFrame, Any, None]:
+        return ckan.datastore_search_generator(resource_id=resource_id, **file_query, search_all=True)
 
 
 class RequestFileMapperABC(RequestMapperABC, ABC):
@@ -161,8 +157,8 @@ class RequestFileMapperLimit(RequestFileMapperABC):
         row_count = ckan.datastore_search_row_count(resource_id)
         return [{"offset": self.limit*counter, "limit": self.limit} for counter in range(row_count // self.limit + 1)]
 
-    def download_file_query(self, ckan: CkanApi, resource_id: str, file_query:dict) -> pd.DataFrame:
-        return ckan.datastore_search(resource_id=resource_id, offset=file_query["offset"], limit=file_query["limit"], search_all=True)
+    def download_file_query(self, ckan: CkanApi, resource_id: str, file_query:dict) -> Generator[pd.DataFrame, Any, None]:
+        return ckan.datastore_search_generator(resource_id=resource_id, offset=file_query["offset"], limit=file_query["limit"], search_all=True)
 
 
 class RequestFileMapperIndexKeys(RequestFileMapperABC):